# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estacionamento API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Lista todos os veículos",
                "operationId": "727634867f59796290a2fff276979198",
                "responses": {
                    "200": {
                        "description": "Lista de veículos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetAllVehiclesResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vehicle": {
            "post": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Cria um novo veículo",
                "operationId": "d923ccc5f5f836fbd8001d15ec79edc1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "licensePlate": {
                                        "description": "Placa do veículo",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Veículo criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateVehicleResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vehicle/{id}": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Obtém um veículo por ID",
                "operationId": "16eeb91ddf4438d36ac18d4b0e7f8b35",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do veículo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetVehicleByIdResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Atualiza os detalhes de um veículo",
                "operationId": "8227bc3280cf3c2e9717f931148e0211",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "manufacturer": {
                                        "description": "Fabricante do veículo",
                                        "type": "string"
                                    },
                                    "color": {
                                        "description": "Cor do veículo",
                                        "type": "string"
                                    },
                                    "model": {
                                        "description": "Modelo do veículo",
                                        "type": "string"
                                    },
                                    "licensePlate": {
                                        "description": "Placa do veículo",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Veículo atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UpdateVehicleResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/vehicle/exit": {
            "post": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Registra a saída de um veículo",
                "operationId": "bd5c59b59919bcedfe59fedb096337fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "licensePlate": {
                                        "description": "Placa do veículo",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Saída do veículo registrada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExitVehicleResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "CreateVehicleResource": {
                "title": "CreateVehicleResource",
                "description": "Resource representation for the creation of a vehicle.",
                "properties": {
                    "status": {
                        "description": "Status of the operation. True if successful, false otherwise.",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "Errors encountered during the operation, if any.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "Message indicating the success of the operation, if applicable.",
                        "type": "string"
                    },
                    "vehicle": {
                        "description": "Details of the created vehicle.",
                        "properties": {
                            "id": {
                                "description": "ID of the vehicle",
                                "type": "integer"
                            },
                            "manufacturer": {
                                "description": "Manufacturer of the vehicle",
                                "type": "string"
                            },
                            "color": {
                                "description": "Color of the vehicle",
                                "type": "string"
                            },
                            "model": {
                                "description": "Model of the vehicle",
                                "type": "string"
                            },
                            "licensePlate": {
                                "description": "License plate of the vehicle",
                                "type": "string"
                            },
                            "entryTimes": {
                                "description": "Entry time of the vehicle",
                                "type": "string",
                                "format": "date-time"
                            },
                            "departureTimes": {
                                "description": "Departure time of the vehicle, if applicable",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ExitVehicleResource": {
                "title": "ExitVehicleResource",
                "description": "Resource representation for exiting a vehicle.",
                "properties": {
                    "status": {
                        "description": "Status of the operation. True if successful, false otherwise.",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "Errors encountered during the operation, if any.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "Message indicating the success of the operation, if applicable.",
                        "type": "string"
                    },
                    "vehicle": {
                        "description": "Details of the vehicle that exited.",
                        "properties": {
                            "id": {
                                "description": "ID of the vehicle",
                                "type": "integer"
                            },
                            "manufacturer": {
                                "description": "Manufacturer of the vehicle",
                                "type": "string"
                            },
                            "color": {
                                "description": "Color of the vehicle",
                                "type": "string"
                            },
                            "model": {
                                "description": "Model of the vehicle",
                                "type": "string"
                            },
                            "licensePlate": {
                                "description": "License plate of the vehicle",
                                "type": "string"
                            },
                            "entryTimes": {
                                "description": "Entry time of the vehicle",
                                "type": "string",
                                "format": "date-time"
                            },
                            "departureTimes": {
                                "description": "Departure time of the vehicle, if applicable",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "GetAllVehiclesResource": {
                "title": "GetAllVehiclesResource",
                "description": "Resource representation for retrieving all vehicles.",
                "properties": {
                    "status": {
                        "description": "Status of the operation. True if successful, false otherwise.",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "Errors encountered during the operation, if any.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "Message indicating the success of the operation, if applicable.",
                        "type": "string"
                    },
                    "vehicles": {
                        "description": "List of vehicles found.",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "ID of the vehicle",
                                    "type": "integer"
                                },
                                "manufacturer": {
                                    "description": "Manufacturer of the vehicle",
                                    "type": "string"
                                },
                                "color": {
                                    "description": "Color of the vehicle",
                                    "type": "string"
                                },
                                "model": {
                                    "description": "Model of the vehicle",
                                    "type": "string"
                                },
                                "licensePlate": {
                                    "description": "License plate of the vehicle",
                                    "type": "string"
                                },
                                "entryTimes": {
                                    "description": "Entry time of the vehicle",
                                    "type": "string",
                                    "format": "date-time"
                                },
                                "departureTimes": {
                                    "description": "Departure time of the vehicle, if applicable",
                                    "type": "string",
                                    "format": "date-time"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "GetVehicleByIdResource": {
                "title": "GetVehicleByIdResource",
                "description": "Resource representation for retrieving a vehicle by ID.",
                "properties": {
                    "status": {
                        "description": "Status of the operation. True if successful, false otherwise.",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "Errors encountered during the operation, if any.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "Message indicating the success of the operation, if applicable.",
                        "type": "string"
                    },
                    "vehicle": {
                        "description": "Details of the found vehicle.",
                        "properties": {
                            "id": {
                                "description": "ID of the vehicle",
                                "type": "integer"
                            },
                            "manufacturer": {
                                "description": "Manufacturer of the vehicle",
                                "type": "string"
                            },
                            "color": {
                                "description": "Color of the vehicle",
                                "type": "string"
                            },
                            "model": {
                                "description": "Model of the vehicle",
                                "type": "string"
                            },
                            "licensePlate": {
                                "description": "License plate of the vehicle",
                                "type": "string"
                            },
                            "entryTimes": {
                                "description": "Entry time of the vehicle",
                                "type": "string",
                                "format": "date-time"
                            },
                            "departureTimes": {
                                "description": "Departure time of the vehicle, if applicable",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UpdateVehicleResource": {
                "title": "UpdateVehicleResource",
                "description": "Resource representation for updating a vehicle.",
                "properties": {
                    "status": {
                        "description": "Status of the operation. True if successful, false otherwise.",
                        "type": "boolean"
                    },
                    "errors": {
                        "description": "Errors encountered during the operation, if any.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "message": {
                        "description": "Message indicating the success of the operation, if applicable.",
                        "type": "string"
                    },
                    "vehicle": {
                        "description": "Details of the updated vehicle.",
                        "properties": {
                            "id": {
                                "description": "ID of the vehicle",
                                "type": "integer"
                            },
                            "manufacturer": {
                                "description": "Manufacturer of the vehicle",
                                "type": "string"
                            },
                            "color": {
                                "description": "Color of the vehicle",
                                "type": "string"
                            },
                            "model": {
                                "description": "Model of the vehicle",
                                "type": "string"
                            },
                            "licensePlate": {
                                "description": "License plate of the vehicle",
                                "type": "string"
                            },
                            "entryTimes": {
                                "description": "Entry time of the vehicle",
                                "type": "string",
                                "format": "date-time"
                            },
                            "departureTimes": {
                                "description": "Departure time of the vehicle, if applicable",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT token to authenticate"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}